+++ CHANGELOG.md
@@ -1,6 +1,12 @@
+## 1.2.0
+
+- Added funtion to get update details
+- Added function to get download progress with flexible update
+
 ## 1.1.0
 
 **Breaking Changes:**
+
 - Added support for new architecture
 - Removed steps for native setup
 - Added method for Javascript/Typescript with update type (immediate, flexible)